refactor: 레파지토리 내 실패 테스트 삭제 및 서비스 내 실패 테스트 추가
