+++ src/test/java/com/part3/team07/sb01deokhugamteam07/repository/CommentRepositoryCustomImplTest.java
@@ -220,25 +220,7 @@ class CommentRepositoryCustomImplTest {
   }
 
   @Test
-  @DisplayName("댓글 목록 조회 실패 - 커서값이 정렬 조건과 다를경우")
-  void findCommentsFailByInvalidCursor() {
-    String invalidCursor = "not-a-datetime";
-
-    //when & then
-    assertThatThrownBy(() -> commentRepositoryCustom.findCommentByCursor(
-        testReview,
-        "DESC",
-        invalidCursor, // 잘못된 커서
-        null,
-        10,
-        "createdAt"
-    ))
-        .isInstanceOf(InvalidCommentQueryException.class)
-        .hasMessageContaining("잘못된 커서 포맷입니다.");
-  }
-
-  @Test
-  @DisplayName("댓글 목록 조회 실패 - 정렬 조건이 잘못된 입력값일 경우")
+  @DisplayName("댓글 목록 조회 실패 - 정렬 조건이 선택사항에 없을 경우")
   void findCommentsFailByInvalidOrderBy() {
     //given
     String invalidSortBy = "invalidField";
@@ -254,25 +236,6 @@ class CommentRepositoryCustomImplTest {
     ))
         .isInstanceOf(InvalidCommentQueryException.class)
         .hasMessageContaining("지원하지 않는 정렬 필드입니다.");
-  }
-
-  @Test
-  @DisplayName("댓글 목록 조회 실패 - 잘못된 정렬 방향")
-  void findCommentsFailByInvalidDirection() {
-    //given
-    String invalidDirection = "DDD";
-
-    //when & then
-    assertThatThrownBy(() -> commentRepositoryCustom.findCommentByCursor(
-        testReview,
-        invalidDirection,
-        null,
-        null,
-        10,
-        "CreatedAt"
-    ))
-        .isInstanceOf(InvalidCommentQueryException.class)
-        .hasMessageContaining("정렬 방향은 ASC 또는 DESC만 가능합니다.");
   }
 
 }
