+++ src/test/java/com/part3/team07/sb01deokhugamteam07/service/CommentServiceTest.java
@@ -19,6 +19,8 @@ import com.part3.team07.sb01deokhugamteam07.entity.Review;
 import com.part3.team07.sb01deokhugamteam07.entity.User;
 import com.part3.team07.sb01deokhugamteam07.exception.comment.CommentNotFoundException;
 import com.part3.team07.sb01deokhugamteam07.exception.comment.CommentUnauthorizedException;
+import com.part3.team07.sb01deokhugamteam07.exception.comment.InvalidCommentQueryException;
+import com.part3.team07.sb01deokhugamteam07.exception.review.ReviewNotFoundException;
 import com.part3.team07.sb01deokhugamteam07.exception.user.UserNotFoundException;
 import com.part3.team07.sb01deokhugamteam07.mapper.CommentMapper;
 import com.part3.team07.sb01deokhugamteam07.repository.CommentRepository;
@@ -28,7 +30,6 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
@@ -173,7 +174,7 @@ class CommentServiceTest {
 
     //when & then
     assertThatThrownBy(() -> commentService.create(createRequest))
-        .isInstanceOf(NoSuchElementException.class); // 예외 추가 시 변경 예정
+        .isInstanceOf(ReviewNotFoundException.class);
 
   }
 
@@ -415,7 +416,7 @@ class CommentServiceTest {
 
 
   @Test
-  @DisplayName("리뷰에 달린 댓글 목록 조회 성공 - hasNext: true")
+  @DisplayName("댓글 목록 조회 성공 - hasNext: true")
   void findCommentsByReviewId_hasNextTrue() {
     //given
     LocalDateTime time = LocalDateTime.of(2025, 4, 25, 12, 0);
@@ -484,7 +485,7 @@ class CommentServiceTest {
   }
 
   @Test
-  @DisplayName("리뷰에 달린 댓글 목록 조회 성공 - hasNext: False")
+  @DisplayName("댓글 목록 조회 성공 - hasNext: False")
   void findCommentsByReviewId_hasNextFalse() {
     //given
     LocalDateTime time = LocalDateTime.of(2025, 4, 25, 12, 0);
@@ -549,7 +550,7 @@ class CommentServiceTest {
   }
 
   @Test
-  @DisplayName("리뷰에 달린 댓글 목록 조회 실패 - 리뷰 존재X")
+  @DisplayName("댓글 목록 조회 실패 - 리뷰 존재X")
   void findCommentsByReviewId_fail_reviewNotFound() {
     //given
     given(reviewRepository.findById(eq(reviewId))).willReturn(Optional.empty());
@@ -560,6 +561,43 @@ class CommentServiceTest {
         null,
         null,
         3
-    )).isInstanceOf(NoSuchElementException.class); //todo 예외 추가 시 변경 예정
+    )).isInstanceOf(ReviewNotFoundException.class);
   }
+
+  @Test
+  @DisplayName("댓글 목록 조회 실패 - 커서값이 정렬 조건과 다를경우")
+  void findCommentsFailByInvalidCursor() {
+    //given
+    String invalidCursor = "not-a-datetime";
+
+    //when & then
+    assertThatThrownBy(() -> commentService.findCommentsByReviewId(
+        reviewId,
+        "DESC",
+        invalidCursor,
+        null,
+        10
+    ))
+        .isInstanceOf(InvalidCommentQueryException.class)
+        .hasMessageContaining("잘못된 커서 포맷입니다.");
+  }
+
+  @Test
+  @DisplayName("댓글 목록 조회 실패 - 잘못된 정렬 방향")
+  void findCommentsFailByInvalidDirection() {
+    //given
+    String invalidDirection = "DDD";
+
+    //when & then
+    assertThatThrownBy(() -> commentService.findCommentsByReviewId(
+        reviewId,
+        invalidDirection,
+        null,
+        null,
+        10
+    ))
+        .isInstanceOf(InvalidCommentQueryException.class)
+        .hasMessageContaining("정렬 방향은 ASC 또는 DESC만 가능합니다.");
+  }
+
 }
